More build script rearranging

+++ build.sh
@@ -19,26 +19,26 @@ echo "3. Other "
 read -p "1-3?" choice
 
 # Enter custom Team ID and Team Name
-if [ "$choice" == 3  ]
+if [ "$choice" == 3 ]
 then
-echo "Your Team ID and Team Name must match exactly with the signing identity in your keychain."
+  echo "Your Team ID and Team Name must match exactly with the signing identity in your keychain."
 
-read -p "Enter Team ID: " CUSTOMTEAMID
-read -p "Enter Team Name: " CUSTOMTEAMNAME
-fi
+  read -p "Enter Team ID: " CUSTOMTEAMID
+  read -p "Enter Team Name: " CUSTOMTEAMNAME
 
-# Simple Team ID Validation. Apple Team ID always consists of 10 characters
-if  ! [ "${#CUSTOMTEAMID}" == 10  ]
-then
-echo "Error: Team ID is not valid"
-exit 1
+  # Simple Team ID Validation. Apple Team ID always consists of 10 characters
+  if  ! [ "${#CUSTOMTEAMID}" == 10 ]
+  then
+    echo "Error: Team ID is not valid"
+    exit 1
+  fi
 fi
 
 case "$choice" in
   1 ) TEAMID="ZYJ4TZX4UU"; SIGNINGIDENTITY="Developer ID Application: SURFnet B.V. ($TEAMID)";;
   2 ) TEAMID="E85CT7ZDJC"; SIGNINGIDENTITY="Developer ID Application: Egeniq ($TEAMID)";;
   3 ) TEAMID="$CUSTOMTEAMID"; SIGNINGIDENTITY="Developer ID Application: $CUSTOMTEAMNAME ($TEAMID)";;
-  * ) echo "Invalid response."; exit 0;;
+  * ) echo "Invalid response."; exit 1;;
 esac
 
 BRANCH=$(git rev-parse --abbrev-ref HEAD)
@@ -47,9 +47,9 @@ echo "You are currently on branch $BRANCH."
 
 if [[ $BRANCH != "release/"* ]]
 then
-    echo ""
-    echo "You must always build from a release branch. Switch to the correct branch or ask the developer to create it for you."
-    exit
+  echo ""
+  echo "You must always build from a release branch. Switch to the correct branch or ask the developer to create it for you."
+  exit
 fi
 
 VERSION=$(git rev-parse --abbrev-ref HEAD | cut -d "/" -f 2)
@@ -59,7 +59,7 @@ read -p "Continue building $PRODUCT version $VERSION (using $SIGNINGIDENTITY) (y
 case "$choice" in
   y|Y ) ;;
   n|N ) exit 0;;
-  * ) echo "Invalid response."; exit 0;;
+  * ) echo "Invalid response."; exit 1;;
 esac
 
 FILENAME="$TARGET-$VERSION"
@@ -70,13 +70,13 @@ echo "$(tput setaf 2)Bootstrapping dependencies using carthage$(tput sgr 0)"
 # Check if Carthage is installed
 if ! [ -x "$(command -v carthage)" ]; then
   echo 'Carthage is not installed. Installing Carthage' >&2
-  
+
   # Check if Homebrew is installed
   if ! [ -x "$(command -v brew)" ]; then
     echo 'Error: Homebrew is not installed. Install Homebrew manually please from https://brew.sh' >&2
     exit 1
   fi
-  
+
   brew install carthage
 fi
 
@@ -105,39 +105,68 @@ UP=$(find $FILENAME -name "*.up.*.sh" -print)
 codesign -f -s "$SIGNINGIDENTITY" "$UP"
 
 echo ""
-echo "$(tput setaf 2)Creating a disk image$(tput sgr 0)"
+echo "Which distribution format do you want to use?"
+echo "1. None (recommended for local usage)"
+echo "2. create-dmg (recommended for 3rd party distribution)"
+echo "3. DropDMG (recomended for eduVPN distribution) "
+read -p "1-3?" choice
 
-# Check if create-dmg is installed
-if ! [ -x "$(command -v create-dmg)" ]; then
+if [ "$choice" == 1 ]
+then
+  echo ""
+  echo "Done!"
+fi
+
+if [ "$choice" == 2 ]
+then
+  # Check if create-dmg is installed
+  if ! [ -x "$(command -v create-dmg)" ]; then
     # Check if HomeBrew is installed
     if ! [ -x "$(command -v brew)" ]; then
       echo 'Error: Homebrew is not installed. Install Homebrew manually please from https://brew.sh' >&2
       exit 1
     fi
-    
-  brew install create-dmg
+
+    brew install create-dmg
+  fi
+
+  INSTALLERFILENAME="$FILENAME.dmg"
+
+  echo ""
+  echo "$(tput setaf 2)Creating a disk image$(tput sgr 0)"
+
+  create-dmg \
+  --volname "$TARGET" \
+  --window-pos 490 350 \
+  --window-size 490 350 \
+  --icon-size 100 \
+  --icon "$PRODUCT" 100 155 \
+  --hide-extension "$PRODUCT" \
+  --app-drop-link 370 155 \
+   $INSTALLERFILENAME \
+  $FILENAME"/$PRODUCT"
+  codesign -f -s "$SIGNINGIDENTITY" $INSTALLERFILENAME
+
+  echo ""
+  echo "Done!"
 fi
 
-INSTALLERFILENAME="$TARGET-Installer$(date +"%Y-%m-%d-%H:%M:%S.")dmg"
+if [ "$choice" == 3 ]
+then
+  echo ""
+  echo "$(tput setaf 2)Creating a disk image$(tput sgr 0)"
+  # The configuration eduVPN can be used for all products
+  echo "Using: dropdmg --config-name \"eduVPN\" --signing-identity=\"$SIGNINGIDENTITY\" \"$FILENAME/$PRODUCT\""
+  dropdmg --config-name "eduVPN" --signing-identity="$SIGNINGIDENTITY" "$FILENAME/$PRODUCT"
 
-create-dmg \
---volname "$TARGET" \
---volicon "icon-$TARGET.icns" \
---background "background-$TARGET.png" \
---window-pos 490 350 \
---window-size 490 350 \
---icon-size 100 \
---icon "$PRODUCT" 100 155 \
---hide-extension "$PRODUCT" \
---app-drop-link 370 155 \
- $INSTALLERFILENAME \
-$FILENAME"/$PRODUCT"
-codesign -f -s "$SIGNINGIDENTITY" $INSTALLERFILENAME
+  echo ""
+  echo "$(tput setaf 2)Creating app cast XML$(tput sgr 0)"
+  DISTRIBUTIONPATH="../eduvpn-macos-distrib"
+  # Assumptions are being made about the location of this script
+  # Also, this often fails due to extended attribute
+  echo "Using: $DISTRIBUTIONPATH/generate_appcast $DISTRIBUTIONPATH/dsa_priv.pem $DISTRIBUTIONPATH/updates/"
+  $DISTRIBUTIONPATH/generate_appcast $DISTRIBUTIONPATH/dsa_priv.pem $DISTRIBUTIONPATH/updates/
 
-echo ""
-echo "Done! You can now upload the files in the updates folders to your file server. Also remember to merge the release branch into master and tag it."
-
-
-
-
-
+  echo ""
+  echo "Done! You can now upload the files in the updates folders to your file server. Also remember to merge the release branch into master and tag it."
+fi
